Expose public interpret method.
