+++ src/main.rs
@@ -6,8 +6,7 @@ use std::process;
 
 extern crate librlox;
 extern crate parcel;
-use librlox::interpreter::expression::ExpressionInterpreter;
-use librlox::interpreter::Interpreter;
+use librlox::interpreter::expression::interpret;
 use librlox::parser::expression_parser::expression;
 use librlox::scanner;
 use parcel::Parser;
@@ -65,9 +64,7 @@ fn run(source: String) -> ParseResult<usize> {
         .collect();
 
     match expression().parse(&tokens) {
-        Ok(parcel::MatchStatus::Match((_, expr))) => {
-            println!("{}", ExpressionInterpreter::new().interpret(expr).unwrap())
-        }
+        Ok(parcel::MatchStatus::Match((_, expr))) => println!("{}", interpret(expr).unwrap()),
         Ok(parcel::MatchStatus::NoMatch(_)) => println!("No match found"),
         Err(e) => println!("{:?}", e),
     };
